svn merge -r 27229:27471 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@27480 260f80e4-7a28-3924-810f-c04153c831b5

+++ tests/standalone/io/directory_chdir_test.dart
@@ -26,8 +26,15 @@ testChangeDirectory() {
     Directory.current = "..";
     Expect.isTrue(new File("111").existsSync());
     Expect.isTrue(new File("222/333").existsSync());
+    // Deleting the current working directory causes an error.
+    // On Windows, the deletion fails, and on non-Windows, the getter fails.
+    Expect.throws(() {
+      temp.deleteSync(recursive: true);
+      Directory.current;
+    }, (e) => e is DirectoryException);
     Directory.current = initialCurrent;
-    temp.deleteSync(recursive: true);
+    Directory.current;
+    if (temp.existsSync()) temp.deleteSync(recursive: true);
     asyncEnd();
   });
 }
